+++ NEWS.md
@@ -1,8 +1,24 @@
+# renv2docker 0.0.0.9004
+
+Templates
+
+- Added dev dependency: `rappster/renvx`
+- Added templating functionality
+
+--------------------------------------------------------------------------------
+
+# renv2docker 0.0.0.9003
+
+- Added package dependency: `here`
+- Changed `Filesystem` to `Local`
+
+--------------------------------------------------------------------------------
+
 # renv2docker 0.0.0.9002
 
 Renamed functions & devtools
 
-- Renamed
+- Renamed functions
 - Added package dependency: `devtools`
 
 --------------------------------------------------------------------------------